Code for the Pong Game
So it took me a little bit to get the game code up, but here it is! Try to go and understand the logic in the code. To help in the learning process, you guys should take a look at the code and try to add features! Try to add music playing in the background by using the loadSound() method call in the Resources class! Even better, try to add a DeadState class that extends State! Also, perhaps try to see how you can use an array or arraylist (or whatever data structure you feel will work best) to keep track of the amount of times the ball goes off screen. 

There is a lot you guys can do with this code, but the only way you will really learn, is if you go and try to mess around with the code.